respa_exchange: Use resource given for exchange sync
Current version have bug when trying to save new email address the
exchange integration is trying to sync with the old information. This
is because the sync_from_exchange function is doing query to get the
resource, but when saving the resource the sync is called in the models
clean method. That makes it so that the new given email is not used,
but the old that is in the database.