Change Oauth2.0 Exchange authentication to NTLM
We haven't found a way to configure OAuth on the new cloud O365 without
violating security policies that the cloud caretaker has in place.

All OAuth configurations that would technically work with a server-to-server
integration like Respa has require too wide permissions to the Respa
credentials. This is why we go back to NTLM and rethink our approach.

This reverts commit ef4b9d4f3183f706adc73ff218a59eb95fd46f08.