Change NTLM Exchange authentication to Basic Auth
NTLM authentication is not available in the new Office 365 cloud, so we
need to replace it.

Microsoft was going to disable Basic Auth in the cloud, but COVID got
them postpone that plan. Their plan is to still disable Basic Auth,
but a bit later, sometime during 2021.

So this change puts us on a borrowed time and we need a different
long-term plan. However this gets us alive in kicking in production
for now.