Don't use IPv6 for Exchange even if it is available
The cloud O365 Exchange has an allowlist of IP addresses that are
allowed to use "legacy authentication" (such as the Basic Auth we are
using). That allowlist only supports IPv4 addresses. Normally requests
uses IPv6 if it detects that the infrastucture supports IPv6, which our
production environment does. So currently our requests in production are
getting blocked by that allowlist because we are sending requests from
an IPv6 address.

This change attempts to force our requests to always use IPv4, even when
IPv6 would be available. Unfortunately there is no another environment
where IPv6 would be supported, so this is completely untested and based
on hearsay that it might work.