Implement CPU Ceepos payment provider
Add support for CPU's Ceepos payment solution. Ceepos is used by the
City of Tampere. The implementation is very similar to the Bambora
provider which has been used as an example for the Ceepos provider
implementation, where applicable.

The only change to the Respa core is the addition of an exception class
and handling of it in the payments API. Other than that there shouldn't
be any conflicts with Helsinki when updating from upstream in the future.

Refs TAM-65